+++ src/coords.rs
@@ -1,9 +1,11 @@
 //! Screen and pixel coordinate systems.
 
+use std::fmt::Display;
+
 use eframe::egui::{Pos2, Rect};
 
 /// Integer point, e.g. pixel coordinates.
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Point {
     pub x: i32,
     pub y: i32,
@@ -12,6 +14,12 @@ pub struct Point {
 impl Point {
     pub fn new(x: i32, y: i32) -> Self {
         Self { x, y }
+    }
+}
+
+impl Display for Point {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "({}, {})", self.x, self.y)
     }
 }