+++ src/update_area.rs
@@ -2,6 +2,8 @@ use std::collections::HashMap;
 
 use bit_vec::BitVec;
 
+use crate::{coords::Point, line};
+
 /// Pixels or cells that are affected by an update
 pub struct UpdateArea {
     pixels: Vec<(i32, i32)>,
@@ -13,6 +15,16 @@ impl UpdateArea {
             pixels: vec![(x, y)],
         }
     }
+
+    /// Create an UpdateArea from a line between two pixels.
+    /// To avoid overdrawing the ending point of a previous line,
+    /// the starting pixel `p0` is not included in the line.
+    pub fn pixel_line(p0: Point, p1: Point) -> Self {
+        UpdateArea {
+            pixels: line::line(p0, p1).skip(1).map(|p| (p.x, p.y)).collect(),
+        }
+    }
+
     pub fn iter_pixels(&self) -> impl Iterator<Item = (i32, i32)> + '_ {
         self.pixels.iter().copied()
     }