+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.10.0] - 2022-01-04
 ### Added
 - "Replace Color" tool to replace one color with another.
 - "Swap Colors" tool to replace one color with another, and vice versa.
@@ -64,7 +66,8 @@ This makes it impossible to try to draw with colors 8-15 unless in a multicolor 
 
 
 <!-- next-url -->
-[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/vilcans/pixel_pen/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/vilcans/pixel_pen/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/vilcans/pixel_pen/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/vilcans/pixel_pen/compare/v0.6.0...v0.7.0